Create SQL Image & Persistent Volume Claim
SQL Server does  not work on the new Macbook:

The original SQL Server for Linux image has not yet been
ported to the ARM64 architecture, which is what the  M1 Mac
uses.

Therefore, it is not possible to use the below image:
mcr.microsoft.com/mssql/server

Instead, I need to use:
mcr.microsoft.com/azure-sql-edge

At present the password is exposed. Need to retry by storing
the password as a secret.